+++ packages/eyes-common/lib/config/BrowserType.js
@@ -10,6 +10,8 @@ const BrowserType = {
   IE_11: 'ie',
   IE_10: 'ie10',
   EDGE: 'edge',
+  EDGE_CHROMIUM: 'edgechromium',
+  EDGE_LEGACY: 'edgelegacy',
   SAFARI: 'safari',
   CHROME_ONE_VERSION_BACK: 'chrome-one-version-back',
   CHROME_TWO_VERSIONS_BACK: 'chrome-two-versions-back',
@@ -17,6 +19,7 @@ const BrowserType = {
   FIREFOX_TWO_VERSIONS_BACK: 'firefox-two-versions-back',
   SAFARI_ONE_VERSION_BACK: 'safari-one-version-back',
   SAFARI_TWO_VERSIONS_BACK: 'safari-two-versions-back',
+  EDGE_CHROMIUM_ONE_VERSION_BACK: 'edgechromium-one-version-back',
 }
 
 Object.freeze(BrowserType)